# openapi-spec.yaml
openapi: 3.0.3
info:
  title: 智慧乡村综合服务平台 API
  description: |
    智慧乡村综合服务平台的完整 API 规范，涵盖村民管理、村务治理、信息公示、生活服务四大核心模块。

    ## 架构说明
    - **主API服务器** (端口3001): 监控、稳定性管理、多语言支持、通知服务
    - **村务服务器** (端口5000): 核心村务管理、Socket.IO实时通信、紧急广播

    ## 核心功能模块
    1. **村委管理**: 村委信息管理、村务协同、任务调度
    2. **村民管理**: 信息数字化、在线办事、档案管理
    3. **村务治理**: 财务管理、项目管理、公开透明
    4. **信息公示**: 政策公告、村务通知、关键词查询
    5. **安全防护**: 数据加密、权限管控、操作审计
  version: 2.0.0
  contact:
    name: 智慧乡村平台团队
    email: support@smartvillage.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: 主API服务器 (监控、通知、多语言)
  - url: http://localhost:5000
    description: 村务服务器 (核心服务、Socket.IO)
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - 认证管理
      summary: 用户登录
      description: 支持多种登录方式：账号密码、手机验证码、人脸识别
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                loginType:
                  type: string
                  enum: [password, sms, face]
                  description: 登录方式
                username:
                  type: string
                  description: 用户名或手机号
                password:
                  type: string
                  description: 密码（password方式必需）
                smsCode:
                  type: string
                  description: 短信验证码（sms方式必需）
                faceData:
                  type: string
                  description: 人脸识别数据（face方式必需）
              required:
                - loginType
                - username
            examples:
              password_login:
                summary: 密码登录
                value:
                  loginType: "password"
                  username: "admin"
                  password: "password123"
              sms_login:
                summary: 短信验证码登录
                value:
                  loginType: "sms"
                  username: "13800138000"
                  smsCode: "123456"
              face_login:
                summary: 人脸识别登录
                value:
                  loginType: "face"
                  username: "13800138000"
                  faceData: "base64_encoded_face_data"
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/residents:
    get:
      tags:
        - 村民管理
      summary: 获取村民档案列表
      description: 分页获取村民档案，支持多条件筛选和隐私保护
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          schema:
            type: string
          description: 姓名搜索（模糊匹配）
        - name: familyType
          in: query
          schema:
            type: string
            enum: [normal, poverty, single, disabled, veteran]
          description: 家庭类型筛选
        - name: ageRange
          in: query
          schema:
            type: string
            pattern: '^\d+-\d+$'
          description: 年龄范围（如：18-60）
      responses:
        '200':
          description: 村民档案列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResidentProfile'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          description: 无权限访问

    post:
      tags:
        - 村民管理
      summary: 新建村民档案
      description: 创建新村民档案，自动生成一户一码
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResidentCreate'
      responses:
        '201':
          description: 档案创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ResidentProfile'
                  householdQRCode:
                    type: string
                    description: 一户一码二维码URL
...
  /api/announcements:
    get:
      tags:
        - 信息公示
      summary: 获取公告列表
      description: 获取政策公告和村务通知，支持关键词搜索
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [policy, notice, activity, emergency]
          description: 公告类别
        - name: keyword
          in: query
          schema:
            type: string
          description: 关键词搜索
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, normal, high, urgent]
          description: 优先级筛选
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: 公告列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Announcement'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - 信息公示
      summary: 发布新公告
      description: 发布政策公告或村务通知
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncementCreate'
      responses:
        '201':
          description: 公告发布成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
...
    Pagination:
      type: object
      properties:
        current:
          type: integer
          description: 当前页码
        pageSize:
          type: integer
          description: 每页大小
        total:
          type: integer
          description: 总记录数
        totalPages:
          type: integer
          description: 总页数

    # ==================== 认证相关 ====================
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
          description: JWT访问令牌
        refreshToken:
          type: string
          description: 刷新令牌
        user:
          type: object
          properties:
            id:
              type: string
            username:
              type: string
            name:
              type: string
            role:
              type: string
            village:
              type: string
            permissions:
              type: array
              items:
                type: string
        expiresIn:
          type: integer
          description: 令牌有效期（秒）
      required:
        - success
        - token
        - user
...
    ResidentProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          description: 姓名
        phone:
          type: string
          description: 联系电话（脱敏）
        idCard:
          type: string
          description: 身份证号（脱敏）
        address:
          type: string
        familyType:
          type: string
          enum: [normal, poverty, single, disabled, veteran]
        householdSize:
          type: integer
        registrationDate:
          type: string
          format: date
        lastUpdated:
          type: string
          format: date-time
        qrCode:
          type: string
          description: 一户一码
...
    ResidentCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 20
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
        idCard:
          type: string
        address:
          type: string
        familyType:
          type: string
          enum: [normal, poverty, single, disabled, veteran]
        familyMembers:
          type: array
          items:
            $ref: '#/components/schemas/FamilyMemberCreate'
      required:
        - name
        - phone
        - idCard
        - address
...
    FamilyMemberCreate:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
          enum: [spouse, parent, child, sibling, grandparent, grandchild, other]
        age:
          type: integer
        idCard:
          type: string
        phone:
          type: string
        occupation:
          type: string
      required:
        - name
        - relationship
        - age

    # ==================== 村务治理 ====================
...
    Announcement:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [policy, notice, activity, emergency]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        publishDate:
          type: string
          format: date-time
        expiryDate:
          type: string
          format: date-time
        author:
          type: string
        attachments:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              url:
                type: string
              size:
                type: integer
        views:
          type: integer
        tags:
          type: array
          items:
            type: string
        isTop:
          type: boolean
          description: 是否置顶
...
    AnnouncementCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 200
        content:
          type: string
          minLength: 10
        category:
          type: string
          enum: [policy, notice, activity, emergency]
        priority:
          type: string
          enum: [low, normal, high, urgent]
          default: normal
        expiryDate:
          type: string
          format: date-time
        attachments:
          type: array
          items:
            type: string
            description: 附件文件路径
        tags:
          type: array
          items:
            type: string
        isTop:
          type: boolean
          default: false
        targetAudience:
          type: array
          items:
            type: string
            enum: [all, committee, residents, elderly, youth]
      required:
        - title
        - content
        - category
...
  responses:
...
    Unauthorized:
      description: 未授权访问
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "未授权访问"
              code:
                type: string
                example: "UNAUTHORIZED"
...
    TooManyRequests:
      description: 请求过于频繁
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "请求过于频繁，请稍后再试"
              retryAfter:
                type: integer
                description: 建议重试间隔（秒）
...